Made date and time user input. Added exceptions to client ip address when starting program to connect to server.
The user is now required to enter the day,month,year and time for a meeting before sending the request message. However, we still have to add restrictions to the range of numbers, i.e. past dates, negative number.

The client now has to enter a valid IPAddress to connect to a server.